requirements.txt
Jinja2==2.11.2
napalm==3.1.0
ncclient==0.6.9
netmiko==3.2.0
nornir==3.0.0
nornir-napalm==0.1.1
nornir-netmiko==0.1.1
nornir-utils==0.1.0
pandas==1.1.2
paramiko==2.6.0
PyYAML==5.3.1
requests==2.22.0
virtualenv==20.0.31
xmltodict==0.12.0
